requirements.txt
python-dotenv
pandas 
numpy
pymongo
certifi
pymongo[srv]
scikit-learn
dill
pyaml   
xgboost
catboost
mlflow
dagshub
fastapi
uvicorn
flask
##-e .